+++ client/src/common/commands.ts
@@ -1,11 +1,9 @@
 import * as vscode from "vscode";
 import { LanguageClient } from "vscode-languageclient/node";
 
-import { UserSession } from "./auth/github";
-import { COMMANDS, EXTENSION_ID, EXTENSION_NAME, PUBLISHER, SESSION_USER } from "./constants";
+import { COMMANDS, EXTENSION_ID, EXTENSION_NAME, PUBLISHER } from "./constants";
 
 import { ExceptionHandlingCommandProvider } from './providers/exceptionProvider';
-import { trackExceptionHandlingResultFeedback, trackFeatureUsage, trackUserInterestInCustomRules } from "./logs";
 import { authenticateUser, removeApiKey } from './auth/api';
 
 const WORKBENCH_ACTIONS = {
@@ -30,21 +28,6 @@ export async function registerCommands(
         () => removeApiKey(context, client, () => deactivate(context))
     ));
 
-    const addCustomRuleCommand = vscode.commands.registerCommand(
-        COMMANDS.ADD_CUSTOM_RULE,
-        () => {
-			const storedUser: UserSession = context.globalState.get(SESSION_USER);
-            if (storedUser) {
-                trackUserInterestInCustomRules(storedUser.email || storedUser.github_login);
-            } else {
-                const serverHost = vscode.env.machineId;
-                trackUserInterestInCustomRules(serverHost);
-            }
-            vscode.window.showInformationMessage('Thanks for your interest in automated rules setup. This feature is coming soon! 🚀', "Sounds good");
-        }
-    );
-    context.subscriptions.push(addCustomRuleCommand);
-
     const openWalkthroughCommand = vscode.commands.registerCommand(
         COMMANDS.OPEN_WALKTHROUGH,
         () => vscode.commands.executeCommand(WORKBENCH_ACTIONS.OPEN_WALKTHROUGH, `${EXTENSION_ID}.gettingStarted`, true)
@@ -65,8 +48,6 @@ export async function registerCommands(
                 vscode.window.showErrorMessage('Please sign in to use this feature.');
                 return;
             }
-
-            trackFeatureUsage(token as string, 'analyzeExceptionHandling');
 
             const editor = vscode.window.activeTextEditor;
 
@@ -94,19 +75,4 @@ export async function registerCommands(
             await commandProvider.provideExceptionHandling(document, functionName, lineNumber);
         })
     );
-
-    context.subscriptions.push(vscode.commands.registerCommand(
-        COMMANDS.NPLUSONE_FEEDBACK,
-        async (uri: vscode.Uri, diagnostic: vscode.Diagnostic) => {
-            const feedback = await vscode.window.showQuickPick(['Good', 'Bad', 'False Positive'], {
-                placeHolder: 'How would you rate this N+1 suggestion?'
-            });
-
-            if (feedback) {
-                vscode.window.showInformationMessage('Thank you for your response!');
-                const token = context.globalState.get(COMMANDS.USER_API_KEY) || "Anonymous";
-                trackExceptionHandlingResultFeedback(token as string, feedback);
-            }
-        }
-    ));
 }