+++ scripts/e2e.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
 echo "Running Client Tests..."
-export CODE_TESTS_PATH="$(pwd)/client/out/test"
+export CODE_TESTS_PATH="$(pwd)/client/out/client/src/test"
 export CODE_TESTS_WORKSPACE="$(pwd)/client/testFixture"
-node "$(pwd)/client/out/test/runTest"
+node "$(pwd)/client/out/client/src/test/runTest"
 
 # Check if the client tests passed before running the server tests
 if [ $? -ne 0 ]; then